+++ sdk/servicebus/azure-servicebus/samples/sync_samples/sample_code_servicebus.py
@@ -85,6 +85,17 @@ def example_create_servicebus_sender_sync():
     with servicebus_client:
         queue_sender = servicebus_client.get_queue_sender(queue_name=queue_name)
     # [END create_servicebus_sender_from_sb_client_sync]
+
+    # [START create_topic_sender_from_sb_client_sync]
+    import os
+    from azure.servicebus import ServiceBusClient
+    servicebus_connection_str = os.environ['SERVICE_BUS_CONNECTION_STR']
+    topic_name = os.environ['SERVICE_BUS_TOPIC_NAME']
+    servicebus_client = ServiceBusClient.from_connection_string(conn_str=servicebus_connection_str)
+    with servicebus_client:
+        queue_sender = servicebus_client.get_topic_sender(topic_name=topic_name)
+    # [END create_topic_sender_from_sb_client_sync]
+
     return queue_sender